Inventory order records now store counts as corrections
Before, it was treated the same as receiving a purchase, which would
have lead to the counts being doubled after the second inventory.
Now it stores detail records for correction info.  If the counts are
correct for an item, then there's no correction - no detail record

Inventory::Command::ItemTransaction::execute() checks the return value
from apply_barcode_to_order and bails out if it returns false